Define each pin as a type unsigned integer of length 8 bits

+++ bike.ino
@@ -1,31 +1,31 @@
 #define JOYSTICKSHIELD true
 
 #ifdef JOYSTICKSHIELD
-  #define UP_BTN          2
-  #define RIGHT_BTN       3
-  #define DOWN_BTN        4
-  #define LEFT_BTN        5
-  #define E_BTN           6
-  #define F_BTN           7
-  #define JOYSTICK_BTN    8
-  #define pinHallSensor   13
-  #define JOYSTICK_AXIS_X A0
-  #define JOYSTICK_AXIS_Y A1
+  #define UP_BTN          (uint8_t)2
+  #define RIGHT_BTN       (uint8_t)3
+  #define DOWN_BTN        (uint8_t)4
+  #define LEFT_BTN        (uint8_t)5
+  #define E_BTN           (uint8_t)6
+  #define F_BTN           (uint8_t)7
+  #define JOYSTICK_BTN    (uint8_t)8
+  #define pinHallSensor   (uint8_t)13
+  #define JOYSTICK_AXIS_X (uint8_t)A0
+  #define JOYSTICK_AXIS_Y (uint8_t)A1
   int buttons[] = {UP_BTN, DOWN_BTN, LEFT_BTN, RIGHT_BTN, E_BTN, F_BTN, JOYSTICK_BTN};
 #else
-  #define pinInterruptMPU 0
-  #define pinButtonB      1
+  #define pinInterruptMPU (uint8_t)0
+  #define pinButtonB      (uint8_t)1
   // Pins 2 and 3 and for SCA and SCL of the MPU6050
-  #define pinButtonA      4
-  #define pinButtonAdd    5
-  #define pinButtonR      6
-  #define pinButtonL      7
-  #define pinHallSensor   8
-  #define pinSPTDLeftJS   9
-  #define pinRotationLED  12
-  #define pinSpeedLED     13
-  #define pintLAnalogX    A0
-  #define pintLAnalogY    A1
+  #define pinButtonA      (uint8_t)4
+  #define pinButtonAdd    (uint8_t)5
+  #define pinButtonR      (uint8_t)6
+  #define pinButtonL      (uint8_t)7
+  #define pinHallSensor   (uint8_t)8
+  #define pinSPTDLeftJS   (uint8_t)9
+  #define pinRotationLED  (uint8_t)12
+  #define pinSpeedLED     (uint8_t)13
+  #define pintLAnalogX    (uint8_t)A0
+  #define pintLAnalogY    (uint8_t)A1
 #endif
 
 
@@ -39,7 +39,7 @@ SwitchJoystick_ Joystick;
 boolean gameStarted = false;
 
 void setup() {
-  #ifdef JOYSTICKSHIELD 
+  #ifdef JOYSTICKSHIELD
     for (int i; i < 7; i++) pinMode(buttons[i], INPUT_PULLUP);
   #else
     pinMode(pinButtonL, INPUT);
@@ -47,11 +47,11 @@ void setup() {
     pinMode(pinButtonAdd, INPUT);
     pinMode(pinButtonA, INPUT);
     pinMode(pinButtonB, INPUT);
-  
+
     pinMode(pintLAnalogX, INPUT);
     pinMode(pintLAnalogY, INPUT);
     pinMode(pinSPTDLeftJS, INPUT);
-    
+
     pinMode(pinSpeedLED, OUTPUT);
     pinMode(pinRotationLED, OUTPUT);
 
@@ -78,7 +78,7 @@ void loop() {
       digitalWrite(pinSpeedLED, LOW);
       digitalWrite(pinRotationLED, HIGH);
     #endif
-    
+
     delay(100);
     #ifdef JOYSTICKSHIELD
       gameStarted = (digitalRead(RIGHT_BTN) == LOW);
